refactor: refactor attachment download logic to use a dedicated utility function
